Changed the length column in C2CSEGMENT from smallint to int. This shows the dangers of blindly copying the structure of an existing table. A smallint might be adequate for a read-to-contig mapping, but it is disastrously inadequate for a contig-to-contig mapping where segments can easily exceed the size of a smallint.
git-svn-id: svn+ssh://svn.internal.sanger.ac.uk/repos/svn/arcturus/trunk@636 69345dfe-1762-0410-a7d4-b9dc18918749

+++ arcturus/utils/create-tables.sql
@@ -467,7 +467,7 @@ CREATE TABLE C2CSEGMENT (
   mapping_id mediumint(8) unsigned NOT NULL,
   cstart int(10) unsigned NOT NULL default '0',
   pstart int(10) unsigned NOT NULL default '0',
-  length smallint(5) unsigned NOT NULL default '1',
+  length int(10) unsigned NOT NULL default '1',
   KEY mapping_id (mapping_id)
 ) TYPE=MyISAM;
